requirements-dev.txt
pytest>=8.0
pytest-mock>=3.12
pyyaml>=6.0
